+++ R/hydro_imgw_annual.R
@@ -1,7 +1,7 @@
 #' Semi-annual and annual hydrological data
 #'
 #' Downloading hydrological data for the semi-annual and annual period
-#' available in the danepubliczne.imgw.pl collection
+#' available in the dane.imgw.pl collection
 #'
 #' @param year vector of years (e.g., 1966:2000)
 #' @param coords add coordinates of the stations (logical value TRUE or FALSE)
@@ -84,7 +84,7 @@ hydro_imgw_annual =  function(year, coords = FALSE, value = "H", station = NULL,
   #station selection
   if (!is.null(station)) {
     if (is.character(station)) {
-      all_data = all_data[all_data$`Nazwa stacji` %in% station, ]
+      all_data = all_data[substr(all_data$`Nazwa stacji`,1,nchar(station))==station, ]
       if (nrow(all_data) == 0){
         stop("Selected station(s) is not available in the database.", call. = FALSE)
       }