+++ R/hydro_imgw_monthly.R
@@ -1,6 +1,6 @@
 #' Monthly hydrological data
 #'
-#' Downloading monthly hydrological data from the danepubliczne.imgw.pl collection
+#' Downloading monthly hydrological data from the dane.imgw.pl collection
 #'
 #' @param year vector of years (e.g., 1966:2000)
 #' @param coords add coordinates of the stations (logical value TRUE or FALSE)
@@ -83,7 +83,7 @@ hydro_imgw_monthly <- function(year, coords = FALSE, station = NULL, col_names= 
   #station selection
   if (!is.null(station)) {
     if (is.character(station)) {
-      all_data <- all_data[all_data$`Nazwa stacji` %in% station, ]
+      all_data = all_data[substr(all_data$`Nazwa stacji`,1,nchar(station))==station, ]
       if (nrow(all_data) == 0){
         stop("Selected station(s) is not available in the database.", call. = FALSE)
       }