+++ R/meteo_imgw.R
@@ -1,6 +1,6 @@
 #' Meteorological data from IMGW
 #'
-#' Downloading hourly, daily, and monthly meteorological data from the SYNOP / CLIMATE / PRECIP stations available in the danepubliczne.imgw.pl collection
+#' Downloading hourly, daily, and monthly meteorological data from the SYNOP / CLIMATE / PRECIP stations available in the dane.imgw.pl collection
 #'
 #' @param interval temporal resolution of the data ("hourly", "daily", "monthly")
 #' @param rank rank of the stations: "synop" (default), "climate" or "precip"
@@ -9,7 +9,7 @@
 #' @param coords add coordinates of the station (logical value TRUE or FALSE)
 #' @param col_names three types of column names possible: "short" - default, values with shorten names, "full" - full English description, "polish" - original names in the dataset
 #' @param ... other parameters that may be passed to the 'shortening' function that shortens column names
-#' @param station vector of hydrological stations danepubliczne.imgw.pl can be name of station CAPITAL LETTERS(character)
+#' @param station vector of hydrological stations dane.imgw.pl can be name of station CAPITAL LETTERS(character)
 #' It accepts names (characters in CAPITAL LETTERS) or stations' IDs (numeric)
 #' @export
 #' @return A data.frame with columns describing the meteorological parameters (e.g. temperature, wind speed, precipitation) where each row represent a measurement,