+++ R/sounding_wyoming.R
@@ -63,25 +63,35 @@ sounding_wyoming <- function(wmo_id, yy, mm, dd, hh){
 
   temp <- tempfile()
   test_url(url, temp)
-  #download.file(url, temp)
-
-  txt <- read.fwf(file = temp, widths = 1000)
-  sects <- grep(pattern = "PRE>", x = txt$V1)
-  if (length(sects) == 0){
-    stop("HTTP status was '503 Service Unavailable'. Have you provided a correct station id? Please check wmo_id numbers at
-          https://ogimet.com/display_stations.php?lang=en&tipo=AND&isyn=&oaci=&nombre=&estado=&Send=Send")
-  }
-  df <- read.fwf(file = temp, skip = sects[1] + 4, widths = rep(7, 11),
-                   n = (sects[2] - (sects[1] + 5)))
   
-  colnames(df) <- c("PRES", "HGHT", "TEMP", "DWPT", "RELH",
-                    "MIXR", "DRCT", "SKNT", "THTA", "THTE", "THTV")
+  # run only if downloaded file is valid
+  df = NULL
+  if(!is.na(file.size(temp)) & (file.size(temp) > 800)) { 
 
-  txt <- read.fwf(file = temp, skip = sects[2] + 1, widths = 1000,
-                    n = (sects[3] - (sects[2] + 2)), stringsAsFactors = FALSE)$V1
-  df2 <- as.data.frame(matrix(data = unlist(strsplit(txt, split = ": ")), ncol = 2, byrow = TRUE))
-  colnames(df2) <- c("parameter"," value")
-  df <- list(df, df2)
+    txt <- read.fwf(file = temp, widths = 1000)
+    sects <- grep(pattern = "PRE>", x = txt$V1)
+    if (length(sects) == 0){
+      stop("HTTP status was '503 Service Unavailable'. Have you provided a correct station id?
+      Please check wmo_id numbers at:
+      http://weather.uwyo.edu/upperair/sounding.html")
+    }
+    df <- read.fwf(file = temp, skip = sects[1] + 4, widths = rep(7, 11),
+                     n = (sects[2] - (sects[1] + 5)))
+    
+    colnames(df) <- c("PRES", "HGHT", "TEMP", "DWPT", "RELH",
+                      "MIXR", "DRCT", "SKNT", "THTA", "THTE", "THTV")
+  
+    txt <- read.fwf(file = temp, skip = sects[2] + 1, widths = 1000,
+                      n = (sects[3] - (sects[2] + 2)), stringsAsFactors = FALSE)$V1
+    df2 <- as.data.frame(matrix(data = unlist(strsplit(txt, split = ": ")), ncol = 2, byrow = TRUE))
+    colnames(df2) <- c("parameter"," value")
+    df <- list(df, df2)
+  
+    } else { # end of checking file size / problems with internet connection
+     cat(paste0("Service not working or wmo_id or date not correct. Check url:\n", url)) 
+    }
+      
+  
   unlink(temp)
 
   return(df)