+++ NEWS.md
@@ -1,3 +1,22 @@
+# climate 0.9.9
+
+* Changing URL `danepubliczne.imgw.pl` to `dane.imgw.pl` where needed
+* Fixing minor ogimet and IMGW bugs
+* Do not stop downloading data from `ogimet.com`, instead check for all available data in given period of time
+
+# climate 0.9.8
+
+* Adding informative message if problems with NOAA hourly dataset occur
+* Informative message if problems with downloading detected for non-IMGW dataset
+
+# climate 0.9.7
+
+* stop working if no internet connection detected
+
+# climate 0.9.6
+
+* Adding nearest_stations_noaa for NOAA hourly dataset
+
 # climate 0.9.5
 
 * Following CRAN policies